DuAn1/DuAn1/src/Entity/HoaDonChiTiet.java

package Entity;


public class HoaDonChiTiet {
    private int maHDCT;
    private int maHD;
    private String maSP;
    private int soLuong;
    private double Gia;
    private String ghiChu;

    public int getMaHD() {
        return maHD;
    }

    public void setMaHD(int maHD) {
        this.maHD = maHD;
    }

    public String getMaSP() {
        return maSP;
    }

    public void setMaSP(String maSP) {
        this.maSP = maSP;
    }

    public int getSoLuong() {
        return soLuong;
    }

    public void setSoLuong(int soLuong) {
        this.soLuong = soLuong;
    }

    public double getGia() {
        return Gia;
    }

    public void setGia(double Gia) {
        this.Gia = Gia;
    }

    public int getMaHDCT() {
        return maHDCT;
    }

    public void setMaHDCT(int maHDCT) {
        this.maHDCT = maHDCT;
    }

    public String getGhiChu() {
        return ghiChu;
    }

    public void setGhiChu(String ghiChu) {
        this.ghiChu = ghiChu;
    }

    

    public HoaDonChiTiet() {
    }

    public HoaDonChiTiet(int maHDCT, int maHD, String maSP, int soLuong, double Gia, String ghiChu) {
        this.maHDCT = maHDCT;
        this.maHD = maHD;
        this.maSP = maSP;
        this.soLuong = soLuong;
        this.Gia = Gia;
        this.ghiChu = ghiChu;
    }

    
    
}
